// types/applicationInfo.ts
export type applicationInfo = {
  id: string;
  imageName: string;
  firstName: string;
  lastName: string;
  age: string;
  phone: string;
  gender: string;
  address: string;
  city: string;
  major: string;
  skillsandabilities: string;
  references: string;
  experiances: string;
  cv: string;
  appliedPosition: string;
};